# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Descripción de este servidor backend",
        "title": "API Backend de la aplicación besigabi",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1/",
    "paths": {
...
        "/api/v1/besigabi/creacionTurnos": {
            "get": {
                "description": "Descripción detallada del endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreacionTurnos"
                ],
                "summary": "Resumen de lo que hace  el endpoint",
                "parameters": [
                    {
                        "description": "Parámetros de la consulta de las ordenes iniciales",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CreacionTurnosRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CreacionTurnosResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un turno en el sistema validando cupos y límites por módulo y origen de forma atómica",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreacionTurnos"
                ],
                "summary": "Crea un turno",
                "parameters": [
                    {
                        "description": "Parámetros para la creación del turno",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CrearTurnoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CrearTurnoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/besigabi/creacionTurnos/cargarConfigLIS": {
            "get": {
                "description": "Recupera el separador de muestras desde la tabla ENT4028",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreacionTurnos"
                ],
                "summary": "Obtiene la configuración de LIS",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.ConfigLISResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/besigabi/creacionTurnos/companias": {
            "post": {
                "description": "Consulta las tablas ENT5802 y ENT5814 para obtener las compañías asociadas a una sede",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreacionTurnos"
                ],
                "summary": "Obtiene el listado de compañías",
                "parameters": [
                    {
                        "description": "Parámetros de la consulta de compañías",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CompaniaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "description": "Listado de compañías",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CompaniaResponse"
                            }
                        }
                    },
                    "204": {
                        "description": "Sin Data"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/besigabi/creacionTurnos/confirmarConfigSedes": {
            "get": {
                "description": "Verifica si se manejan sedes y si no, retorna la sede por defecto",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreacionTurnos"
                ],
                "summary": "Confirmar configuración de sedes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    },
                    "204": {
                        "description": "No se encontró sede por Default"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/besigabi/creacionTurnos/modulos": {
            "post": {
                "description": "Consulta la tabla ENT5818 para obtener los módulos por sede",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreacionTurnos"
                ],
                "summary": "Obtiene el listado de módulos",
                "parameters": [
                    {
                        "description": "Parámetros de la consulta de módulos",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.ModuloRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "description": "Listado de módulos",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.ModuloResponse"
                            }
                        }
                    },
                    "204": {
                        "description": "Sin Data"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/besigabi/creacionTurnos/sedes": {
            "get": {
                "description": "Consulta la tabla ENT0021 para obtener las sedes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreacionTurnos"
                ],
                "summary": "Obtiene el listado de sedes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "description": "Listado de sedes",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.SedeResponse"
                            }
                        }
                    },
                    "204": {
                        "description": "Sin Data"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/besigabi/creacionTurnos/tipoDocumento": {
            "get": {
                "description": "Consulta la tabla ENT0024 para obtener los tipos de documento activos (C05 = 1)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreacionTurnos"
                ],
                "summary": "Obtiene el listado de tipos de documento",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "description": "Listado de tipos de documento",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.TipoDocumentoResponse"
                            }
                        }
                    },
                    "204": {
                        "description": "Sin Data"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v1/besigabi/creacionTurnos/tipoServicio": {
            "post": {
                "description": "Consulta las tablas ENT5800 y ENT5815 para obtener los tipos de servicio",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreacionTurnos"
                ],
                "summary": "Obtiene el listado de tipos de servicio",
                "parameters": [
                    {
                        "description": "Parámetros de la consulta de tipos de servicio",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.TipoServicioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "description": "Listado de tipos de servicio",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.TipoServicioResponse"
                            }
                        }
                    },
                    "204": {
                        "description": "Sin Data"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/besigabi/creacionTurnos/tipoTurno": {
            "post": {
                "description": "Consulta las tablas ENT5816 y ENT5810 para obtener los tipos de turno por servicio",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreacionTurnos"
                ],
                "summary": "Obtiene el listado de tipos de turno",
                "parameters": [
                    {
                        "description": "Parámetros de la consulta de tipos de turno",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.TipoTurnoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "description": "Listado de tipos de turno",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.TipoTurnoResponse"
                            }
                        }
                    },
                    "204": {
                        "description": "Sin Data"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/besigabi/creacionTurnos/verificarConfigCompanias": {
            "get": {
                "description": "Consulta la tabla ENT5803 para verificar si se manejan empresas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreacionTurnos"
                ],
                "summary": "Verifica la configuración de compañías",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.ConfigCompaniasResponse"
                        }
                    },
                    "204": {
                        "description": "Sin Configuración"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CompaniaRequest": {
            "description": "Estructura que representa un request para obtener compañías",
            "type": "object",
            "properties": {
                "idSede": {
                    "type": "integer",
                    "example": -1
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CompaniaResponse": {
            "description": "Estructura que representa una compañía",
            "type": "object",
            "properties": {
                "codigoCompania": {
                    "type": "string",
                    "example": "C01"
                },
                "idCompania": {
                    "type": "integer",
                    "example": 1
                },
                "nombreCompania": {
                    "type": "string",
                    "example": "Compañía de Prueba"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.ConfigCompaniasResponse": {
            "description": "Estructura que representa la configuración de compañías",
            "type": "object",
            "properties": {
                "manejaEmpresas": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.ConfigLISResponse": {
            "description": "Estructura que representa la configuración de LIS",
            "type": "object",
            "properties": {
                "separadorMuestra": {
                    "type": "string",
                    "example": "|"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CreacionTurnosRequest": {
            "description": "Estructura que representa un request de un endpoint de la API",
            "type": "object",
            "properties": {
                "creacionTurnos": {
                    "description": "Incluir descripción del campo",
                    "type": "string",
                    "example": "'Valor_de_ejemplo'"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CreacionTurnosResponse": {
            "description": "Estructura que representa un response de un endpoint de la API",
            "type": "object",
            "properties": {
                "creacionTurnos": {
                    "description": "Incluir descripción del campo",
                    "type": "string",
                    "example": "'Valor_de_ejemplo'"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CrearTurnoRequest": {
            "type": "object",
            "properties": {
                "CodigoModulo": {
                    "type": "string"
                },
                "NombreModulo": {
                    "type": "string"
                },
                "accion": {
                    "type": "string"
                },
                "idCompania": {
                    "type": "integer"
                },
                "idSede": {
                    "type": "integer"
                },
                "idServicio": {
                    "type": "integer"
                },
                "idTipoTurno": {
                    "type": "integer"
                },
                "origen": {
                    "type": "integer"
                },
                "paciente": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.PacienteRequest"
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CrearTurnoResponse": {
            "type": "object",
            "properties": {
                "Accion": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.AccionResponse"
                },
                "Compania": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.CompaniaResponse"
                },
                "Sede": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.SedeResponse"
                },
                "Servicio": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.TipoServicioResponse"
                },
                "Turno": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.DatosTurnoResponse"
                },
                "paciente": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.PacienteRequest"
                },
                "tipoTurno": {
                    "$ref": "#/definitions/develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.TipoTurnoResponse"
                }
            }
        },
...
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.ModuloRequest": {
            "description": "Estructura que representa un request para obtener módulos",
            "type": "object",
            "properties": {
                "idSede": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.ModuloResponse": {
            "description": "Estructura que representa un módulo",
            "type": "object",
            "properties": {
                "CodigoModulo": {
                    "type": "string",
                    "example": "200"
                },
                "IdSede": {
                    "type": "integer",
                    "example": 1
                },
                "NombreModulo": {
                    "type": "string",
                    "example": "Modulo de Prueba"
                }
            }
        },
...
        "develop_private_CLTech_besigabi_internal_api_creacionTurnos_domain.SedeResponse": {
            "description": "Estructura que representa una sede",
            "type": "object",
            "properties": {
                "codSede": {
                    "type": "string",
                    "example": "S01"
                },
                "idSede": {
                    "type": "integer",
                    "example": 1
                },
                "nomSede": {
                    "type": "string",
                    "example": "Sede Central"
                }
            }
        },
...
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
    }
}